Move database to repository module
