+++ app/src/main/java/com/bhaktaprogram/notehelper/App.kt
@@ -1,0 +1,13 @@
+package com.bhaktaprogram.notehelper
+
+import android.app.Application
+import com.bhaktaprogram.notehelper.repository.di.DaggerDatabaseComponent
+
+class App : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+
+        DaggerDatabaseComponent.builder().setContext(this).build()
+    }
+}
