+++ repository/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.library'
     id 'kotlin-android'
+    id 'kotlin-kapt'
 }
 
 apply from: "$rootDir/deps_versions.gradle"
@@ -38,6 +39,13 @@ dependencies {
 
     implementation kotlin
     implementation core_ktx
+
+    implementation room.runtime
+    implementation room.ktx
+    kapt room.compiler
+
+    implementation dagger.dagger
+    kapt dagger.compiler
 
     testImplementation test.junit
 }
